+++ src/ir/lubs.cpp
@@ -14,6 +14,7 @@
  * limitations under the License.
  */
 
+#include "ir/find_all.h"
 #include "ir/lubs.h"
 #include "ir/utils.h"
 #include "wasm-type.h"
